Primera parte asignada a mi, terminada

+++ PalcoNet/Abm Grado/ListGrado.cs
@@ -16,8 +16,8 @@ namespace PalcoNet.Abm_Grado
 {
     public partial class ListGrado : CustomForm
     {
-        Grado grado;
         RepoGradoPublicacion repoGrado = new RepoGradoPublicacion();
+        private BindingSource bindingSource = new BindingSource();
 
         public ListGrado()
         {
@@ -26,80 +26,58 @@ namespace PalcoNet.Abm_Grado
 
         private void btnLimpiar_Click(object sender, EventArgs e)
         {
-            txtNombreGrado.Text = "";
-            numComisionGrado.Value = 0;
+            txtNombreGrado.Clear();
+            txtComisionGrado.Clear();
         }
 
         private void ListGrado_Load(object sender, EventArgs e)
         {
             this.dataGridView1.AllowUserToResizeColumns = false;
             this.dataGridView1.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
-            this.dataGridView1.SelectionMode = DataGridViewSelectionMode.RowHeaderSelect;
-            this.dataGridView1.RowHeaderMouseClick += selectedRowsButton_Click;
+            this.dataGridView1.SelectionMode = DataGridViewSelectionMode.CellSelect;
             this.dataGridView1.MultiSelect = false;
-            this.dataGridView1.SelectionChanged += OnSelectionChanged;
-            
+            this.dataGridView1.RowHeadersVisible = false;
+            this.dataGridView1.AutoGenerateColumns = false;
+            this.columnEliminar.DisplayIndex = this.dataGridView1.Columns.Count - 1;
+            this.columnEditar.DisplayIndex = this.dataGridView1.Columns.Count - 2;
         }
 
         private void btnBuscar_Click(object sender, EventArgs e)
         {
-            Grado grado = new Grado(
-                (int)numComisionGrado.Value, 
-                txtNombreGrado.Text);
-            DataTable table = repoGrado.GetGradosBy(grado);
-            dataGridView1.DataSource = table.DefaultView;
+            bindingSource.Clear();
+            List<Grado> grados = repoGrado.GetGradosBy(txtNombreGrado.Text, txtComisionGrado.Text);
+            grados.ForEach((g) => bindingSource.Add(g));
+            this.dataGridView1.DataSource = bindingSource;
         }
-
-        private void selectedRowsButton_Click(object sender, EventArgs e)
-        {
-            DataGridViewRow row = dataGridView1.SelectedRows[0];
-            grado = buildGrado(row);
-        }
-
-        private void OnSelectionChanged(object sender, EventArgs e)
-        {
-            if (this.dataGridView1.SelectedRows.Count == 1)
-            {
-                this.btnBorrar.Enabled = true;
-                this.btnEditar.Enabled = true;
-            }
-            else
-            {
-                this.btnBorrar.Enabled = false;
-                this.btnEditar.Enabled = false;
-            }
-
-        }
-
 
         private Grado buildGrado(DataGridViewRow row)
         {
             Grado grado = new Grado();
-            grado.setId((int)row.Cells[columnGradoId.Name].Value);
-            grado.SetDescripcion((String)row.Cells[columnGradoDescripcion.Name].Value);
-            grado.SetComision((int)row.Cells[columnGradoComision.Name].Value);
+            grado.Id = (int)row.Cells[columnGradoId.Name].Value;
+            grado.Descripcion = (String)row.Cells[columnGradoDescripcion.Name].Value;
+            grado.Comision = (int)row.Cells[columnGradoComision.Name].Value;
             return grado;
-        }
-
-        private void btnEditar_Click(object sender, EventArgs e)
-        {
-            FormManager.getInstance().Open(new UpdateGrado(grado));
-        }
-
-        private void btnBorrar_Click(object sender, EventArgs e)
-        {
-            DialogResult dialogResult = MessageBox.Show("¿Desea eliminar la fila?", "Eliminar Grado", MessageBoxButtons.YesNo);
-            if (dialogResult == DialogResult.Yes)
-            {
-                DataGridViewRow row = dataGridView1.SelectedRows[0];
-                repoGrado.DeleteGrado(buildGrado(row));
-                this.dataGridView1.Rows.Remove(row);
-            }            
         }
 
         private void dataGridView1_CellContentClick(object sender, DataGridViewCellEventArgs e)
         {
+            DataGridViewRow row = dataGridView1.CurrentRow;
+            Grado grado = buildGrado(row);
 
+            if (e.ColumnIndex == dataGridView1.Columns.IndexOf(this.columnEditar))
+            {
+                Form EditForm = new UpdateGrado(grado);
+                EditForm.ShowDialog();
+            }
+            if (e.ColumnIndex == dataGridView1.Columns.IndexOf(this.columnEliminar))
+            {
+                DialogResult dialogResult = MessageBox.Show("¿Desea eliminar la fila?", "Eliminar Grado", MessageBoxButtons.YesNo);
+                if (dialogResult == DialogResult.Yes)
+                {
+                    repoGrado.DeleteGrado(buildGrado(row));
+                    this.dataGridView1.Rows.Remove(row);
+                }    
+            }
         }
 
     }
